+++ ChangeLog.md
@@ -22,6 +22,10 @@ Other enhancements:
 
 Bug fixes:
 
+* Ignore duplicate files for a single module when a Haskell module was
+  generated from a preprocessor file. See
+  [#4076](https://github.com/commercialhaskell/stack/issues/4076).
+
 
 ## v1.9.0 (release candidate)